% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{as_count_matrix}
\alias{as_count_matrix}
\title{Coerce a count table to a numeric samples x features matrix}
\usage{
as_count_matrix(otu)
}
\description{
Coerce a count table to a numeric samples x features matrix
}
\keyword{internal}
