% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pursue.R
\name{cauchy_combine}
\alias{cauchy_combine}
\title{Cauchy (ACAT) combination of p-values across the columns of a matrix}
\usage{
cauchy_combine(P, weights = NULL)
}
\description{
Cauchy (ACAT) combination of p-values across the columns of a matrix
}
\keyword{internal}
