% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centering.R
\name{center_effects}
\alias{center_effects}
\title{Centre every tested column of the abundance arm}
\usage{
center_effects(arm)
}
\description{
Centre every tested column of the abundance arm
}
\keyword{internal}
