% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centering.R
\name{enull_center}
\alias{enull_center}
\title{Empirical-null centre of a vector of estimates with known SEs}
\usage{
enull_center(b, se, max_iter = 500L, tol = 1e-08)
}
\arguments{
\item{b}{estimates; @param se standard errors (same length); NA allowed.}
}
\value{
list(delta, se_delta, pi0, eta, omega2, w = posterior null probability, n_used)
}
\description{
Empirical-null centre of a vector of estimates with known SEs
}
\keyword{internal}
