% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pursue.R
\name{hetero_hint}
\alias{hetero_hint}
\title{Rough heteroscedasticity screen for a binary tested term}
\usage{
hetero_hint(otu, meta, design)
}
\description{
Rough heteroscedasticity screen for a binary tested term
}
\keyword{internal}
