% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PURSUE-package.R
\docType{package}
\name{PURSUE-package}
\alias{PURSUE}
\alias{PURSUE-package}
\title{PURSUE: two-part differential abundance analysis}
\description{
See \code{\link[=pursue]{pursue()}}.
}
\author{
\strong{Maintainer}: Robertas Tupikas \email{robertast.tupikas@gmail.com}

}
\keyword{internal}
